fixed UUIDs getting re-generated and storage cache sync issues on invalidation

+++ src/main/java/com/raoulvdberge/refinedstorage/network/MessagePortableGridFluidDelta.java
@@ -37,7 +37,7 @@ public class MessagePortableGridFluidDelta
         int size = buf.readInt();
 
         for (int i = 0; i < size; ++i) {
-            long delta = buf.readInt();
+            long delta = buf.readLong();
             clientDeltas.add(Pair.of(StackUtils.readFluidGridStack(buf), delta));
         }
     }
